GaryVoiceDemo/agent.py
import logging

from dotenv import load_dotenv
from livekit.agents import (
    AutoSubscribe,
    JobContext,
    JobProcess,
    WorkerOptions,
    cli,
    llm,
)
from livekit.agents.pipeline import VoicePipelineAgent
from livekit.plugins import silero, openai

load_dotenv(dotenv_path=".env.local")
logger = logging.getLogger("voice-agent")


def prewarm(proc: JobProcess):
    proc.userdata["vad"] = silero.VAD.load()


async def entrypoint(ctx: JobContext):
    initial_ctx = llm.ChatContext().append(
        role="system",
        text=(
            "You are a voice assistant created by Noah Santoni the AI Expert. Your interface with users will be voice. "
            "You should use short and concise responses, and avoiding usage of unpronouncable punctuation. "
            "You were created as a demo to showcase the capabilities of Born and Brand's agents framework."
        ),
    )

    logger.info(f"connecting to room {ctx.room.name}")
    await ctx.connect(auto_subscribe=AutoSubscribe.AUDIO_ONLY)

    # Wait for the first participant to connect
    participant = await ctx.wait_for_participant()
    logger.info(f"starting voice assistant for participant {participant.identity}")

    # Initialize OpenAI TTS plugin with Kokoro settings, but using tts-1 model name
    kokoro_tts = openai.TTS(
        model="tts-1",  # Use tts-1 to be compatible with LiveKit
        voice="am_michael", # Select a voice - see available voices at http://localhost:8880/v1/audio/voices
        api_key="not-needed", #Not needed for local instance
        base_url="http://localhost:8880/v1"  # Point to Kokoro FastAPI endpoint
    )

    agent = VoicePipelineAgent(
        vad=ctx.proc.userdata["vad"],
        stt=openai.STT.with_groq(model="distil-whisper-large-v3-en"),
        llm=openai.LLM.with_groq(model="llama-3.3-70b-versatile"),
        tts=kokoro_tts,  # Use Kokoro TTS (through OpenAI-compatible endpoint)
        chat_ctx=initial_ctx,
    )

    agent.start(ctx.room, participant)

    # The agent should be polite and greet the user when it joins :)
    await agent.say("Hey, how can I help you today?", allow_interruptions=True)


if __name__ == "__main__":
    cli.run_app(
        WorkerOptions(
            entrypoint_fnc=entrypoint,
            prewarm_fnc=prewarm,
        ),
    )

